fix: refactor ipfs api into its own module
