+++ hermes/bin/src/runtime_extensions/hermes/ipfs/host.rs
@@ -1,18 +1,15 @@
 //! IPFS host implementation for WASM runtime.
 
-use super::state::hermes_ipfs_publish;
+use super::state::{
+    hermes_ipfs_add_file, hermes_ipfs_content_validate, hermes_ipfs_evict_peer,
+    hermes_ipfs_get_dht_value, hermes_ipfs_get_file, hermes_ipfs_pin_file, hermes_ipfs_publish,
+    hermes_ipfs_put_dht_value, hermes_ipfs_subscribe,
+};
 use crate::{
     runtime_context::HermesRuntimeContext,
-    runtime_extensions::{
-        bindings::hermes::ipfs::api::{
-            DhtKey, DhtValue, Errno, Host, IpfsContent, IpfsFile, IpfsPath, MessageData, MessageId,
-            PeerId, PubsubTopic,
-        },
-        hermes::ipfs::state::{
-            hermes_ipfs_add_file, hermes_ipfs_content_validate, hermes_ipfs_evict_peer,
-            hermes_ipfs_get_dht_value, hermes_ipfs_get_file, hermes_ipfs_pin_file,
-            hermes_ipfs_put_dht_value, hermes_ipfs_subscribe,
-        },
+    runtime_extensions::bindings::hermes::ipfs::api::{
+        DhtKey, DhtValue, Errno, Host, IpfsContent, IpfsFile, IpfsPath, MessageData, MessageId,
+        PeerId, PubsubTopic,
     },
 };
